fix: focus active agreements heading after id-porten authorization is deleted
